Se organizo mejor el uso de los metodos, features y steps falta los selectores
