+++ features/support/change_story_state_until_restricted_steps.js
@@ -1,25 +1,28 @@
 const { Given, When, Then } = require('@cucumber/cucumber');
-const StoryUtils = require('./story_utils');
+const StoryCreationUtils = require('./story_creation_utils');
+const StoryPointsUtils = require('./story_points_utils');
+const StoryButtonsUtils = require('./story_buttons_utils');
+const StoryVerificationUtils = require('./story_verification_utils');
 
 Given('I have created multiple stories in the backlog for move histories accepted', async function () {
     console.log('I have created multiple stories in the backlog for move histories accepted');
-    await StoryUtils.createMultipleStories(5);
+    await StoryCreationUtils.createMultipleStories(5);
 });
 
 When('I start, finish, deliver, accept, and collapse the story', async function () {
     console.log('Changing the story state from unstarted to accepted and collapsing');
     
-    await StoryUtils.enterStory(); 
-    await StoryUtils.openStoryPointsDropdown();
-    await StoryUtils.selectRandomPoints();   
-    await StoryUtils.pressStartButton();     
-    await StoryUtils.pressFinishButton();  
-    await StoryUtils.pressDeliverButton(); 
-    await StoryUtils.pressAcceptButton();  
-    await StoryUtils.pressCollapseButton();  
+    await StoryButtonsUtils.enterStory();
+    await StoryPointsUtils.openStoryPointsDropdown();
+    await StoryPointsUtils.selectRandomPoints(); 
+    await StoryButtonsUtils.pressStartButton();
+    await StoryButtonsUtils.pressFinishButton();
+    await StoryButtonsUtils.pressDeliverButton();
+    await StoryButtonsUtils.pressAcceptButton();
+    await StoryButtonsUtils.pressCollapseButton();
 });
 
 Then('the accepted story should not be movable within the backlog', async function () {
     console.log('Verifying that the accepted story cannot be moved');
-    await StoryUtils.verifyStoryCannotBeMoved();
+    await StoryVerificationUtils.verifyStoryCannotBeMoved();
 });
