+++ features/support/delete_bug_story_steps.js
@@ -1,15 +1,16 @@
 const { When, Then } = require("@cucumber/cucumber");
-const StoryUtils = require("../../features/support/story_utils");
+const StoryButtonsUtils = require("./story_buttons_utils");  
+const StoryVerificationUtils = require("./story_verification_utils");  
 const chai = require('chai');
 const expect = chai.expect;
 
 When('I delete the bug type story from the backlog', async function() {
     console.log("I delete the bug type story from the backlog");
-    await StoryUtils.deleteBugStory();
+    await StoryButtonsUtils.deleteBugStory();  
 });
 
 Then('I should see the backlog empty', async function() {
     console.log("I should see the backlog empty");
-    const isBacklogEmpty = await StoryUtils.isBacklogEmpty();
+    const isBacklogEmpty = await StoryVerificationUtils.isBacklogEmpty(); 
     expect(isBacklogEmpty).to.be.true;
 });
