+++ features/support/delete_feature_story_steps.js
@@ -1,9 +1,7 @@
-const { When, Then } = require("@cucumber/cucumber");
-const StoryUtils = require("../../features/support/story_utils");
-const chai = require('chai');
-const expect = chai.expect;
+const { When } = require("@cucumber/cucumber");
+const StoryButtonsUtils = require("./story_buttons_utils");
 
 When('I delete the feature type story from the backlog', async function() {
     console.log("I delete the feature type story from the backlog");
-    await StoryUtils.deleteFeatureStory();
+    await StoryButtonsUtils.deleteFeatureStory(); 
 });
